forgot to fork this before making edits...oops
What the f*** are these tables? 1.1, 1.2...caption, etc. :0
The location stuff in lines 190:197 are not true...my genie-matlab directory doesn't have muffingen, just a readme file that says the muffingen.m code is on this github :-)

There were a number of wrong-facing left-hand side 'quotes' that I changed to ` cause I think that works?